Upgrade Menu Mockup
Created a new scene called UpgradeMenu.
Created a new script called MGR_UI.
Created a folder inside the Texture folder called "Render Textures".
Added a folder call "Media" to store random videos and photos.
Create upgrades. Hovering over an upgrades reveals a pop-up with a name, description, and a video of the upgrade.
Press and hold down left click on an upgrade to acquire it.
To Do:
Pop-up display needs to update immediately when player acquires an upgrade.
Create a radial progress meter used when purchasing an upgrade(WIP).