Episode 4 Complete - Multiple Noise Filters
There are now 2 types of noise filters, a simple noise filter and a ridgid noise filter. Both filters can have their own parameters.
Add the ability to render a particular part of the planet (front, back, top, etc) instead of the whole planet. This will make it easier to test parameters in the editor on a higher resolution.